Fix C4100, C4189, C4245, and C4456
* C4100 - Unreferenced function parameter(s)
Microsoft's docs suggest adding a "dummy" reference to the parameter(s) in question, so I did.

* C4189 - Unreferenced local variable
`rapidjson::Value::isArray()` is declared `const`, so eliminating the second call shouldn't affect anything (in theory).

* C4245 - Signed/unsigned mismatch
Explicitly casting to `unsigned int` resolves this, and I'm FAIRLY certain this doesn't change behavior? (I even tested this in Compiler Explorer to be sure 🤣)

* C4456 - Local identifier shadowing
Changing the loop variable from `i` to `j` resolves this. As much as I hate using multiple single-letter variable names, it WAS the most obvious solution.